create ContactForm component
